+++ packages/file_selector/file_selector/example/lib/open_image_page.dart
@@ -3,7 +3,7 @@
 // found in the LICENSE file.
 
 // TODO(stuartmorgan): https://github.com/flutter/flutter/issues/111906
-// ignore_for_file: prefer_const_constructors
+// ignore_for_file: prefer_const_constructors, prefer_const_literals_to_create_immutables
 
 import 'dart:io';