+++ packages/file_selector/file_selector_ios/example/lib/open_image_page.dart
@@ -19,7 +19,11 @@ class OpenImagePage extends StatelessWidget {
     // ignore: prefer_const_constructors
     final XTypeGroup typeGroup = XTypeGroup(
       label: 'images',
+      // TODO(stuartmorgan): https://github.com/flutter/flutter/issues/111906
+      // ignore: prefer_const_literals_to_create_immutables
       extensions: <String>['jpg', 'png'],
+      // TODO(stuartmorgan): https://github.com/flutter/flutter/issues/111906
+      // ignore: prefer_const_literals_to_create_immutables
       macUTIs: <String>['public.image'],
     );
     final XFile? file = await FileSelectorPlatform.instance