+++ packages/file_selector/file_selector_macos/example/lib/open_text_page.dart
@@ -16,6 +16,8 @@ class OpenTextPage extends StatelessWidget {
     // ignore: prefer_const_constructors
     final XTypeGroup typeGroup = XTypeGroup(
       label: 'text',
+      // TODO(stuartmorgan): https://github.com/flutter/flutter/issues/111906
+      // ignore: prefer_const_literals_to_create_immutables
       extensions: <String>['txt', 'json'],
     );
     final XFile? file = await FileSelectorPlatform.instance