+++ script/tool/lib/src/create_all_plugins_app_command.dart
@@ -10,13 +10,13 @@ import 'package:pub_semver/pub_semver.dart';
 import 'package:pubspec_parse/pubspec_parse.dart';
 
 import 'common/core.dart';
-import 'common/plugin_command.dart';
+import 'common/package_command.dart';
 import 'common/repository_package.dart';
 
 const String _outputDirectoryFlag = 'output-dir';
 
 /// A command to create an application that builds all in a single application.
-class CreateAllPluginsAppCommand extends PluginCommand {
+class CreateAllPluginsAppCommand extends PackageCommand {
   /// Creates an instance of the builder command.
   CreateAllPluginsAppCommand(
     Directory packagesDir, {