+++ script/tool/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.11
+
+* Renames `publish-plugin` to `publish`.
+* Renames arguments to `list`:
+    * `--package` now lists top-level packages (previously `--plugin`).
+    * `--package-or-subpackage` now lists top-level packages (previously
+      `--package`).
+
 ## 0.10.0+1
 
 * Recognizes `run_test.sh` as a developer-only file in `version-check`.